# src/main/resources/static/swagger/bonggonge.yaml
openapi: 3.0.0
info:
  title: bonggonge API Console
  version: 2.0.3
servers:
  - url: http://172.27.0.121:8080/
...
paths:
...
  /users/join:
    post:
      tags:
        - User APIs
      summary: 회원정보를 입력받아 회원가입을 진행(이메일 인증완료시)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: gildong
                password:
                  type: string
                  example: passdong
                email:
                  type: string
                  example: Gildong.Hong@hancom.com
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: boolean
                    example: true
        '400':
          description: Client error
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: boolean
                    example: false
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      detail:
                        type: string
                    example:
                      - code: param-003
                        detail: Parameter name mismatch
        '500':
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: boolean
                    example: false
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        example: server-001
                      detail:
                        type: string
                        example: Internal server error
  /users/join/check:
    post:
      tags:
        - User APIs
      summary: username의 중복 확인
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: gildong
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: boolean
                    example: true
        '400':
          description: Client error
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: boolean
                    example: false
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      detail:
                        type: string
                    example:
                      - code: param-003
                        detail: Parameter name mismatch
                      - code: param-006
                        detail: Already userd username
        '500':
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: boolean
                    example: false
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        example: server-001
                      detail:
                        type: string
                        example: Internal server error
...
security:
  - ApiKeyAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: token
...